gitAnnotate.vim: match original commits in blame
When using the "git blame" output, the first column contains the partial
commit SHA1 of when the content of the current line was last modified.

For lines that have not been altered since their original commit, the
partial SHA1 is prefixed with a '^' symbol. Update the regex to account
for the optional '^' symbol.

Signed-off-by: Michael Wookey <michaelwookey@gmail.com>